ml/requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
joblib>=1.3.0
fastapi>=0.115.0
uvicorn>=0.30.0
